Firestore build and testing improvements and documentation (#1113)
