+++ .gitignore
@@ -56,3 +56,23 @@ firestore/integration_test_internal/src/ios/ios_app_framework.mm
 firestore/integration_test_internal/src/ios/ios_firebase_test_framework.mm
 firestore/integration_test_internal/src/ios/.clang-format
 firestore/integration_test_internal/src/android/java/com/google/firebase/example
+firestore/integration_test_internal/Podfile.lock
+firestore/integration_test_internal/Pods
+firestore/integration_test_internal/integration_test.xcworkspace/contents.xcworkspacedata
+firestore/integration_test_internal/integration_test.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
+firestore/integration_test_internal/integration_test.xcodeproj/xcshareddata/xcschemes/integration_test.xcscheme
+download_googletest.py
+external
+src/android/java
+src/android/android_app_framework.cc
+src/android/android_firebase_test_framework.cc
+src/app_framework.h
+src/app_framework.cc
+src/desktop/desktop_app_framework.cc
+src/desktop/desktop_firebase_test_framework.cc
+src/empty.swift
+src/firebase_test_framework.cc
+src/firebase_test_framework.h
+src/ios/.clang-format
+src/ios/ios_app_framework.mm
+src/ios/ios_firebase_test_framework.mm
