+++ firestore/integration_test_internal/src/firestore_integration_test.cc
@@ -29,6 +29,7 @@
 #include "android/util_autoid.h"
 #endif  // !defined(__ANDROID__)
 #include "app_framework.h"
+#include "util/locate_emulator.h"
 
 namespace firebase {
 namespace firestore {
@@ -37,37 +38,6 @@ namespace {
 // name of FirebaseApp to use for bootstrapping data into Firestore. We use a
 // non-default app to avoid data ending up in the cache before tests run.
 static const char* kBootstrapAppName = "bootstrap";
-
-// Set Firestore up to use Firestore Emulator via USE_FIRESTORE_EMULATOR
-void LocateEmulator(Firestore* db) {
-  // Use emulator as long as this env variable is set, regardless its value.
-  if (std::getenv("USE_FIRESTORE_EMULATOR") == nullptr) {
-    LogDebug("Using Firestore Prod for testing.");
-    return;
-  }
-
-#if defined(__ANDROID__)
-  // Special IP to access the hosting OS from Android Emulator.
-  std::string local_host = "10.0.2.2";
-#else
-  std::string local_host = "localhost";
-#endif  // defined(__ANDROID__)
-
-  // Use FIRESTORE_EMULATOR_PORT if it is set to non empty string,
-  // otherwise use the default port.
-  std::string port = std::getenv("FIRESTORE_EMULATOR_PORT")
-                         ? std::getenv("FIRESTORE_EMULATOR_PORT")
-                         : "8080";
-  std::string address =
-      port.empty() ? (local_host + ":8080") : (local_host + ":" + port);
-
-  LogInfo("Using Firestore Emulator (%s) for testing.", address.c_str());
-  auto settings = db->settings();
-  settings.set_host(address);
-  // Emulator does not support ssl yet.
-  settings.set_ssl_enabled(false);
-  db->set_settings(settings);
-}
 
 }  // namespace
 
@@ -163,7 +133,7 @@ Firestore* FirestoreIntegrationTest::TestFirestoreWithProjectId(
   Firestore* db = new Firestore(CreateTestFirestoreInternal(app));
   firestores_[db] = FirestoreInfo(name, std::unique_ptr<Firestore>(db));
 
-  LocateEmulator(db);
+  firestore::LocateEmulator(db);
   return db;
 }
